% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{adjust_diplotype}
\alias{adjust_diplotype}
\title{Assigns more common cyp2d6 diplotype if tie in scoring}
\usage{
adjust_diplotype(df)
}
\arguments{
\item{df}{The dataframe that contains the patientID, the genotypes, and the diplotypes from assign_diplotype()}

}
\value{
A dataframe with updated diplotype calls
}
\description{
The function will check the diplotype and alternate diplotype call and ascertain the more likely diplotype.
}
\examples{
result <- adjust_diplotype(df)
}