% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{all_geno_pheno}
\alias{all_geno_pheno}
\title{Complete Summary Table Function}
\usage{
all_geno_pheno(df, genes, phased = FALSE)
}
\arguments{
\item{df}{The dataframe that contains the genotypes you want to convert}

\item{genes}{The specific CYP genes, as a list of strings}
}
\value{
1. The associated star alleles, metabolizer phenotypes, and activity scores for each genotype, formatted as a dataframe. 2. A message detailing missing genotypes.
}
\description{
Combine the dataframes from assign_diplotype, star_to_activity, and star_to_pheno into one large dataframe containing all the information in one place
}
\examples{
summary1 <- all_geno_pheno(df, c("CYP2C9", "CYP2B6"))
summary2 <- all_geno_pheno(df2, "CYP2C19")
}
