% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{assign_diplotype}
\alias{assign_diplotype}
\title{Assigns Star Allele Values, phased and unphased approach}
\usage{
assign_diplotype(df, genes, phased = FALSE)
}
\arguments{
\item{df}{The dataframe that contains the patientID and the genotypes to convert}

\item{genes}{The specific CYP genes, as a list of strings}

\item{phased}{True/False to indicate if input data is phased or unphased}
}
\value{
A dataframe with added columns containing the corresponding diplotypes for the genotypes
}
\description{
The function will first try to assign star alleles using a phased approach, any genotypes left without a matching diplotype will then undergo an unphased approach for assigning diplotypes.
}
\examples{
result <- assign_diplotype(df, c("CYP2C9", "CYP2B6"), phased = FALSE)
result2 <- assign_diplotype(df, "CYP2C19", phased = TRUE)
}
