% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{convert_list_to_string}
\alias{convert_list_to_string}
\title{Helper function for assign_diplotype()}
\usage{
convert_list_to_string(df)
}
\arguments{
\item{df}{The dataframe to check}
}
\value{
df The dataframe with no list columns
}
\description{
Convert any list columns to characters in order to write results to csv
}
