% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP1A2_Allele_def_2}
\alias{CYP1A2_Allele_def_2}
\title{Allele Definition Table V2 for CYP1A2}
\format{
A data frame with 42 variables and 47 rows
\describe{
    \item{CYP1A2}{CYP1A2 star alleles}
    \item{rsID}{CYP1A2 rs IDs}
    }
}
\source{
{Downloaded from Pharmgkb}
}
\usage{
CYP1A2_Allele_def_2
}
\description{
Contains CYP1A2 information about which variants define star alleles based on the update from Pharmvar in Dec. 2024. Preprocessed to only include star allele and rs ID columns.
}
\examples{
data(CYP1A2_Allele_def_2)       #Lazy loading
}
\keyword{datasets}