% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP2B6_Allele_def}
\alias{CYP2B6_Allele_def}
\title{Allele Definition Table for CYP2B6}
\format{
A data frame with 49 variables and 55 rows
\describe{
    \item{CYP2B6.Allele}{CYP2B6 star alleles}
    \item{rsID}{CYP2B6  rs IDs}
    }
}
\source{
{Downloaded from Pharmgkb}
}
\usage{
CYP2B6_Allele_def
}
\description{
Contains CYP2B6 information about which variants define star alleles. Preprocessed to only include star allele and rs ID columns.
}
\examples{
data(CYP2B6_Allele_def)       #Lazy loading
}
\keyword{datasets}
