% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP2B6_data}
\alias{CYP2B6_data}
\title{Data for CYP2B6 diplotypes}
\format{
A data frame with 3 variables and 16 rows
\describe{
    \item{CYP2B6.Diplotype}{CYP2B6 star allele}
    \item{Activity.Score}{CYP2B6 activity score based on the star allele}
    \item{Phenotype}{CYP2B6 metabolizer phenotype based on the star allele}
    }
}
\source{
{Created in house based on the CYP2B6 alleles tested}
}
\usage{
CYP2B6_data
}
\description{
Contains CYP2B6 diplotypes and corresponding activity scores and metabolizer phenotypes
}
\examples{
data(CYP2B6_data)       #Lazy loading
}
\keyword{datasets}
