% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP2C19_Allele_def}
\alias{CYP2C19_Allele_def}
\title{Allele Definition Table for CYP2C19}
\format{
A data frame with 81 variables and 86 rows
\describe{
    \item{CYP2C19.Allele}{CYP2C9 star alleles}
    \item{rsID}{CYP2C19  rs IDs}
    }
}
\source{
{Downloaded from Pharmgkb}
}
\usage{
CYP2C19_Allele_def
}
\description{
Contains CYP2C19 information about which variants define star alleles. Preprocessed to only include star allele and rs ID columns.
}
\examples{
data(CYP2C19_Allele_def)       #Lazy loading
}
\keyword{datasets}
