% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP2C19_data}
\alias{CYP2C19_data}
\title{Data for CYP2C19 diplotypes}
\format{
A data frame with 3 variables and 23 rows
\describe{
    \item{CYP2C19.Diplotype}{CYP2C19 star allele}
    \item{Activity.Score}{CYP2C19 activity score based on the star allele}
    \item{Phenotype}{CYP2C19 metabolizer phenotype based on the star allele}
    }
}
\source{
{Created in house based on the alleles tested}
}
\usage{
CYP2C19_data
}
\description{
Contains CYP2C19 diplotypes and corresponding activity scores and metabolizer phenotypes
}
\examples{
data(CYP2C19_data)       #Lazy loading
}
\keyword{datasets}
