% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP2C9_data}
\alias{CYP2C9_data}
\title{Data for CYP2C9 diplotypes}
\format{
A data frame with 3 variables and 15 rows
\describe{
    \item{CYP2C9.Diplotype}{CYP2C9 star allele}
    \item{Activity.Score}{CYP2C9 activity score based on the star allele}
    \item{Phenotype}{CYP2C9 metabolizer phenotype based on the star allele}
    }
}
\source{
{Created in house based on the alleles tested}
}
\usage{
CYP2C9_data
}
\description{
Contains CYP2C9 diplotypes and corresponding activity scores and metabolizer phenotypes
}
\examples{
data(CYP2C9_data)       #Lazy loading
}
\keyword{datasets}
