% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP2D6_data}
\alias{CYP2D6_data}
\title{Data for CYP2D6 star alleles}
\format{
A data frame with 3 variables and 125 rows
\describe{
    \item{CYP2D6.Diplotype}{CYP2D6 star allele}
    \item{Activity.Score}{CYP2D6 activity score based on the star allele}
    \item{Phenotype}{CYP2D6 metabolizer phenotype based on the star allele}
    }
}
\source{
{Created in house based on the CYP2D6 tier 1 variant alleles. Also includes: 8, 14, 2A, 36+10, 4.013}
}
\usage{
CYP2D6_data
}
\description{
Contains CYP2D6 star alleles and corresponding activity scores
}
\examples{
data(CYP2D6_data)       #Lazy loading
}
\keyword{datasets}
