% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP3A4_Allele_def}
\alias{CYP3A4_Allele_def}
\title{Allele Definition Table for CYP3A4}
\format{
A data frame with 46 variables and 43 rows
\describe{
    \item{CYP3A4}{CYP3A4 star alleles}
    \item{rsID}{CYP3A4 rs IDs}
    }
}
\source{
{Downloaded from Pharmgkb}
}
\usage{
CYP3A4_Allele_def
}
\description{
Contains CYP3A4 information about which variants define star alleles. Preprocessed to only include star allele and rs ID columns.
}
\examples{
data(CYP3A4_Allele_def)       #Lazy loading
}
\keyword{datasets}
