% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP3A5_Allele_def}
\alias{CYP3A5_Allele_def}
\title{Allele Definition Table for CYP3A5}
\format{
A data frame with 6 variables and 7 rows
\describe{
    \item{CYP3A5.Allele}{CYP3A5 star alleles}
    \item{rsID}{CYP3A5  rs IDs}
    }
}
\source{
{Downloaded from Pharmgkb}
}
\usage{
CYP3A5_Allele_def
}
\description{
Contains CYP3A5 information about which variants define star alleles. Preprocessed to only include star allele and rs ID columns.
}
\examples{
data(CYP3A5_Allele_def)       #Lazy loading
}
\keyword{datasets}
