% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CYP3A5_data}
\alias{CYP3A5_data}
\title{Data for CYP3A5 diplotypes}
\format{
A data frame with 3 variables and 21 rows
\describe{
    \item{CYP3A5.Diplotype}{CYP3A5 star allele}
    \item{Activity.Score}{CYP3A5 activity score based on the star allele}
    \item{Phenotype}{CYP3A5 metabolizer phenotype based on the star allele}
    }
}
\source{
{Downloaded from PharmGKB}
}
\usage{
CYP3A5_data
}
\description{
Contains CYP3A5 diplotypes and corresponding activity scores and metabolizer phenotypes
}
\examples{
data(CYP3A5_data)       #Lazy loading
}
\keyword{datasets}
