% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{fill_empty_cells}
\alias{fill_empty_cells}
\title{Fill empty cells in dataframe}
\usage{
fill_empty_cells(df)
}
\arguments{
\item{df}{A data frame containing genotype data with specific columns. The column names should be snps, for example rs1065852.}
}
\value{
A data frame with all the empty cells assigned the wildtype for that snp.
}
\description{
This function iterates through a dataframe (allele definition table in this use case) and fills the empty cells with the genotypes from the first row (wildtype)
}
\examples{
data_filled <- fill_empty_cells(data_filtered)
}
