% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{filter_no_match}
\alias{filter_no_match}
\title{Filter For No Matching Star Alleles}
\usage{
filter_no_match(df, gene)
}
\arguments{
\item{df}{The dataframe that contains the genotypes and the diplotype column}

\item{genes}{The specific CYP genes, as a list of strings}
}
\value{
A dataframe only containing rows that have "no matching star alleles found" as values in the diplotype column.
}
\description{
This function will filter the dataframe to only include rows where no star allele matches were found. It's used within the assign_diplotype function.
}
\examples{
filtered_df <- filter_no_match(df, c("CYP2C9", "CYP2B6"))
filtered_df2 <- filter_no_match(df, "CYP2C19")
}
