% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{find_diplotype}
\alias{find_diplotype}
\title{Assign Star Allele Genotypes}
\usage{
find_diplotype(df, genes)
}
\arguments{
\item{df}{A data frame containing genotype data with specific columns. Each row represents a patient, and columns represent genotypes for specific SNPs. 
Column names should follow the format "gene_snp", for example, "CYP2D6_rs1065852".}

\item{genes}{The specific CYP genes, as a list of strings}
}
\value{
A data frame with an additional column `gene_diplotype` indicating the assigned star allele genotype and messages indicating missing SNPs and/or genotypes.
}
\description{
This function assigns diplotypes based on genotypes dynamically using an allele definition table. Using the SNPs in the input data, this assigns the corresponding diplotype.
"-" will be treated as an indel. 
Genotype data should have the form "A G", "T T", etc.
}
\examples{
find_diplotype(data, c("CYP2D6", "CYP2C9"))
}
