% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{find_missing_data}
\alias{find_missing_data}
\title{Find Missing Genotypes}
\usage{
find_missing_data(df)
}
\arguments{
\item{df}{A data frame containing genotype data with specific columns. Each row represents a patient, and columns represent genotypes for specific SNPs. 
Column names should follow the format "gene_snp", for example, "CYP2D6_rs1065852".}
}
\value{
A list of character vectors, where each element specifies the patient index (row) and column name (SNP) of missing data (NA).
}
\description{
This function checks the input genotype dataframe for any missing genotype entries. If genotypes are missing, they will be returned in a list specifying which patient index and column contains the missing data.
}
\examples{
find_missing_data(data)

}
