% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{format_genotype}
\alias{format_genotype}
\title{Function to format genotypes}
\usage{
format_genotype(genotype)
}
\arguments{
\item{genotype}{the genotype from the current column and row}
}
\value{
the newly formatted genotype
}
\description{
Checks if genotypes contain a space between the alleles, if not, adds one (i.e AA to A A)
}
