% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{generate_phased_combinations}
\alias{generate_phased_combinations}
\title{Generate phased haplotype combinations}
\usage{
generate_phased_combinations(snp_df, snp_cols)
}
\arguments{
\item{snp_df}{The dataframe that contains the genotypes for the interested snps. As a haplotype, ex. "A" at rs1065852}

\item{snp_cols}{The names of the snps as the columns for the dataframe}
}
\value{
A matrix array containing all possible haplotype combinations.
}
\description{
This function will produce all possible combinations with the genotypes provided.
}
