% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{pie_chart}
\alias{pie_chart}
\title{Pie Chart Phenotype Summary Function}
\usage{
pie_chart(df, genes)
}
\arguments{
\item{df}{The dataframe that contains the metabolizer phenotypes}

\item{genes}{The specific CYP genes, as a list of strings}
}
\value{
A summary pie chart for each input gene
}
\description{
Creates a pie chart to summarize the metabolizer phenotypes for each gene
}
\examples{
CYP2D6_pie_chart <- pie_chart(df, "CYP2D6")
}
