% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{process_cyp1a2}
\alias{process_cyp1a2}
\title{Call CYP1A2 Diplotype}
\usage{
process_cyp1a2(
  genotypes,
  row_idx,
  diplotypes,
  score,
  total_score,
  CYP1A2_name
)
}
\arguments{
\item{genotypes}{the preprocessed genotypes, ready to be assessed as haplotypes}

\item{row_idx}{indicates current row of the dataframe}

\item{diplotypes}{a character vector to store any found diplotype matches}

\item{score}{the current score designated for each found star allele}

\item{total_score}{the combined scores to give the total score associated with a diplotype}

\item{CYP1A2_name}{indicates which nomenclature version to use, either new or old}
}
\value{
A matrix containing any unmatched haplotypes, indicating further processing is required
}
\description{
This function will use index polymorphisms to determine the CYP1A2 diplotype. If no diplotype is found, it will expand the attempt using more of the allele definition table. Used within the find_diplotype() function, not meant to be used individually.
}
