% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{process_cyp2d6}
\alias{process_cyp2d6}
\title{Call CYP2D6 Diplotype}
\usage{
process_cyp2d6(
  genotypes,
  row_idx,
  diplotypes,
  cnv_x9,
  cnv_int6,
  cnv_5utr,
  score,
  total_score
)
}
\arguments{
\item{genotypes}{the preprocessed genotypes, ready to be assessed as haplotypes}

\item{row_idx}{indicates current row of the dataframe}

\item{diplotypes}{a character vector to store any found diplotype matches}

\item{cnv_x9}{copy number information for exon 9}

\item{cnv_int6}{copy number information for intron 6}

\item{cnv_5utr}{copy information for 5 prime utr}

\item{score}{the current score designated for each found star allele}

\item{total_score}{the combined scores to give the total score associated with a diplotype}
}
\value{
A matrix containing any unmatched haplotypes, indicating further processing is required
}
\description{
This function will use index polymorphisms to determine the CYP2D6 diplotype. If no diplotype is found, it will expand the attempt using more of the allele definition table. Used within the find_diplotype() function, not meant to be used individually.
}
