% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{reorder_diplotype}
\alias{reorder_diplotype}
\title{Reorder diplotypes}
\usage{
reorder_diplotype(diplotype)
}
\arguments{
\item{diplotype}{The current diplotype to check}
}
\value{
A reformatted diplotype
}
\description{
The function will reorder the diplotypes to make sure the format is small number/big number
}
