% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_CYP2D6_genotypes}
\alias{sample_CYP2D6_genotypes}
\title{Sample Data with CYP2D6 genotypes}
\format{
A data frame with 15 variables and 4 rows
\describe{
    \item{CYP2D6_CNVInt6}{CYP2D6 snp}
    \item{CYP2D6_CNVx9}{CYP2D6 snp}
    \item{CYP2D6_rs1065852}{CYP2D6 snp}
    \item{CYP2D6_rs1080985}{CYP2D6 snp}
    \item{CYP2D6_rs1135840}{CYP2D6 snp}
    \item{CYP2D6_rs16947}{CYP2D6 snp}
    \item{CYP2D6_rs28371706}{CYP2D6 snp}
    \item{CYP2D6_rs28371725}{CYP2D6 snp}
    \item{CYP2D6_rs35742686}{CYP2D6 snp}
    \item{CYP2D6_rs3892097}{CYP2D6 snp}
    \item{CYP2D6_rs5030655}{CYP2D6 snp}
    \item{CYP2D6_rs5030656}{CYP2D6 snp}
    \item{CYP2D6_rs5030865A}{CYP2D6 snp}
    \item{CYP2D6_rs5030865T}{CYP2D6 snp}
    \item{CYP2D6_rs59421388}{CYP2D6 snp}
    }
}
\source{
{Created in house to be used as an example}
}
\usage{
sample_CYP2D6_genotypes
}
\description{
Contains sample genotypes for 15 CYP2D6 snps
}
\examples{
data(sample_CYP2D6_genotypes)       #Lazy loading
}
\keyword{datasets}
