% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{star_to_activity}
\alias{star_to_activity}
\title{Star Allele to Activity Score Conversion}
\usage{
star_to_activity(df, gene)
}
\arguments{
\item{gene}{The specific CYP gene, as a string}

\item{star_alleles}{The star alleles you want to convert, input as a dataframe with one column containing the star alleles. For CYP2D6, a column with CNV_x9 should be included in input dataframe. If there isn't CNV provided, it will default to 2.}
}
\value{
The associated activity score, appended to the original dataframe input
}
\description{
Convert star allele to activity score
}
\examples{
score1 <- star_to_activity(df, "CYP2C9")
}
