% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genotype_conversion.R
\name{star_to_pheno}
\alias{star_to_pheno}
\title{Star Allele to Phenotype Conversion}
\usage{
star_to_pheno(df, gene)
}
\arguments{
\item{gene}{The specific CYP gene, as a string}

\item{star_alleles}{The star alleles you want to convert, input as a dataframe with one column containing star alleles. For CYP2D6, activity score columns need to be provided (i.e output from star_to_activity()).}
}
\value{
The associated metabolizer phenotype, appended to the original dataframe input
}
\description{
Convert star allele to metabolizer phenotype based on activity scores
}
\examples{
pheno1 <- star_to_pheno(df, "CYP2C9")
}
